Refactored the hub page to use its own jotai Provider to manage hub state. Cleaned up hub components and separated viewing from editing functionality into their own components. Added missing React.Suspense component at top level to fix async atom loading causing infinite loops.
