+++ src/state/global.tsx
@@ -4,7 +4,15 @@ import {Link as Generic} from "grommet-icons/icons";
 import {createBrowserSupabaseClient} from "@supabase/auth-helpers-nextjs";
 import {SupabaseClient} from "@supabase/supabase-js";
 
-import {IconDictionary, LinkType, RegionCatalog, Profile, RegionAssociations, LinkDetails} from "../utils/types";
+import {
+    IconDictionary,
+    LinkType,
+    RegionCatalog,
+    Profile,
+    RegionAssociations,
+    LinkDetails,
+    MemberDetails
+} from "../utils/types";
 import {getEPACatalog, getLinkTypes, getOneEarthCatalog} from "../utils/supabase";
 
 export const dbClientAtom = atom<SupabaseClient>((get) => createBrowserSupabaseClient())
@@ -39,3 +47,4 @@ export const epaCatalogAtom = atom<Promise<RegionCatalog>>(async (get) => {
 
 export const regionAssociationsAtom = atom<RegionAssociations | null>(null)
 export const linkDetailsAtom = atom<Array<LinkDetails>>(new Array<LinkDetails>())
+export const membersAtom = atom<Array<MemberDetails>>(new Array<MemberDetails>())
